[vividus] Rename org.vividus.report.MetadataLogger into org.vividus.log.TestInfoLogger
