+++ vividus/src/main/java/org/vividus/steps/SetupSteps.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2021 the original author or authors.
+ * Copyright 2019-2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,13 +17,13 @@
 package org.vividus.steps;
 
 import org.jbehave.core.annotations.AfterStories;
-import org.vividus.report.MetadataLogger;
+import org.vividus.log.TestInfoLogger;
 
 public class SetupSteps
 {
     @AfterStories
     public void afterStories()
     {
-        MetadataLogger.logEnvironmentMetadata();
+        TestInfoLogger.logEnvironmentMetadata();
     }
 }
